backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
python-dotenv
apscheduler
requests
pytest
pytest-flask
# werkzeug is bundled with flask — no extra install needed
# Production (swap in when deploying):
# psycopg2-binary
# celery[redis]
# redis
